refactor: emum 타입 대문자 수정 및 hook 분리
