+++ src/pages/MainPage/ItemList.tsx
@@ -3,45 +3,32 @@ import { useEffect, useState } from 'react';
 import { Button } from '@/components/common/Button';
 import { RankingGoodsItems } from '@/components/common/GoodsItem/Ranking';
 import { Grid } from '@/components/common/layouts/Grid';
+import type { Item } from '@/pages/MainPage/types';
 
-interface Item {
-  imageSrc: string;
-  subtitle: string;
-  title: string;
-  amount: number;
-}
+import { useVisibility } from './handleVisibility';
+import { sortItems } from './sortItems';
 
 interface ItemListProps {
   items: Item[];
 }
 
 const ItemList = ({ items }: ItemListProps) => {
-  const [sortOrder, setSortOrder] = useState<'asc' | 'desc'>('asc');
-  const [visibleItems, setVisibleItems] = useState(6);
+  const [sortOrder, setSortOrder] = useState<'ASC' | 'DESC'>('ASC');
+  const { visibleItems, handleMore, handleLess } = useVisibility(6);
   const [filteredItems, setFilteredItems] = useState<Item[]>([]);
 
   useEffect(() => {
-    const sortedItems = [...items].sort((a, b) =>
-      sortOrder === 'asc' ? a.amount - b.amount : b.amount - a.amount,
-    );
+    const sortedItems = sortItems(items, sortOrder);
     setFilteredItems(sortedItems);
   }, [items, sortOrder]);
-
-  const handleMore = () => {
-    setVisibleItems((prevVisibleItems) => prevVisibleItems + 6);
-  };
-
-  const handleLess = () => {
-    setVisibleItems(6);
-  };
 
   return (
     <section>
       <h2>실시간 급상승 선물랭킹</h2>
       <div>
-        <select value={sortOrder} onChange={(e) => setSortOrder(e.target.value as 'asc' | 'desc')}>
-          <option value="asc">금액 낮은 순</option>
-          <option value="desc">금액 높은 순</option>
+        <select value={sortOrder} onChange={(e) => setSortOrder(e.target.value as 'ASC' | 'DESC')}>
+          <option value="ASC">금액 낮은 순</option>
+          <option value="DESC">금액 높은 순</option>
         </select>
       </div>
       <Grid columns={{ sm: 4, md: 5, lg: 6 }} gap={20}>
