+++ src/pages/MainPage/getItems.ts
@@ -1,11 +1,6 @@
-interface Item {
-  imageSrc: string;
-  subtitle: string;
-  title: string;
-  amount: number;
-}
+import type { Item } from '@/pages/MainPage/types';
 
-const GetItem = (): Item[] => {
+const getItem = (): Item[] => {
   return Array.from({ length: 30 }, () => ({
     imageSrc: `https://img1.kakaocdn.net/thumb/C320x320@2x.fwebp.q82/?fname=https%3A%2F%2Fst.kakaocdn.net%2Fproduct%2Fgift%2Fproduct%2F20240328114550_e72d7854bfb649c68c6bea6a1b85262b.jpg`,
     subtitle: `오제티캔들`,
@@ -14,4 +9,4 @@ const GetItem = (): Item[] => {
   }));
 };
 
-export default GetItem;
+export default getItem;
