+++ src/pages/MainPage/types.ts
@@ -1,0 +1,7 @@
+export interface Item {
+    imageSrc: string;
+    subtitle: string;
+    title: string;
+    amount: number;
+  }
+
